created reservationList component and moved code from Dashboard component
