+++ front-end/src/layout/Routes.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { Redirect, Route, Switch } from "react-router-dom";
-import NewReservation from "../NewReservation";
+import Reservations from "../Reservations";
 import Dashboard from "../dashboard/Dashboard";
 import NotFound from "./NotFound";
 import { today } from "../utils/date-time";
@@ -26,7 +26,7 @@ function Routes() {
         <Dashboard date={today()} />
       </Route>
       <Route path="/reservations/new">
-        <NewReservation />
+        <Reservations />
       </Route>
       <Route>
         <NotFound />
